+++ tests/thrift/test_rw.py
@@ -18,18 +18,42 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-from __future__ import absolute_import
+import pytest
 
-import os
-import sys
-
-from .. import thrift
+from tchannel import thrift
 
 
-base = os.path.dirname(__file__)
-meta = thrift.load(os.path.join(base, 'meta.thrift'))
-sys.modules[__name__ + '.meta'] = meta
+@pytest.mark.parametrize('args, kwargs', [
+    (['tests/data/idls/ThriftTest.thrift'], {}),
+    (['server', 'tests/data/idls/ThriftTest.thrift'], {}),
+    (['tests/data/idls/ThriftTest.thrift', 'server'], {}),
+    ([], {'path': 'tests/data/idls/ThriftTest.thrift'}),
+
+])
+def test_load_backwards_compatibility(args, kwargs):
+    """We initially required a ``service`` argument to always be provided."""
+    assert thrift.load(*args, **kwargs).ThriftTest
 
 
-def health(request):
-    return meta.HealthStatus(ok=True)
+def test_thriftrw_client_without_service_or_hostport_specified():
+    client = thrift.load(
+        'tests/data/idls/ThriftTest.thrift',
+    )
+
+    with pytest.raises(ValueError):
+        client.ThriftTest.testVoid()
+
+
+@pytest.mark.parametrize('hostport, service', [
+    ('127.0.0.1:1234', None),
+    ('127.0.0.1:1234', 'foo'),
+    (None, 'foo'),
+])
+def test_thriftrw_client_with_service_or_hostport_specified(hostport, service):
+    client = thrift.load(
+        'tests/data/idls/ThriftTest.thrift',
+        hostport=hostport,
+        service=service,
+    )
+
+    assert client.ThriftTest.testVoid()